refactor: rename to Loader.vue
