+++ src/platform/assets/components/ActiveMediaAssetCard.vue
@@ -24,18 +24,14 @@
         class="absolute inset-0 flex items-center justify-center bg-modal-card-placeholder-background"
       >
         <!-- Spinner for queued/initialization states -->
-        <LoadingSpinner
-          v-if="isQueued"
-          size="md"
-          class="text-muted-foreground"
-        />
+        <Loader v-if="isQueued" size="md" class="text-muted-foreground" />
         <!-- Error icon for failed state -->
         <i
           v-else-if="isFailed"
           class="icon-[lucide--circle-alert] size-8 text-red-500"
         />
         <!-- Spinner for running without preview -->
-        <LoadingSpinner v-else size="md" class="text-muted-foreground" />
+        <Loader v-else size="md" class="text-muted-foreground" />
       </div>
       <!-- Cancel/Delete button overlay -->
       <Button
@@ -78,7 +74,7 @@
 import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 
-import LoadingSpinner from '@/components/common/LoadingSpinner.vue'
+import Loader from '@/components/common/Loader.vue'
 import Button from '@/components/ui/button/Button.vue'
 import { useJobActions } from '@/composables/queue/useJobActions'
 import type { JobListItem } from '@/composables/queue/useJobList'
