+++ src/renderer/extensions/linearMode/LinearControls.vue
@@ -5,7 +5,7 @@ import { storeToRefs } from 'pinia'
 import { computed, ref, shallowRef } from 'vue'
 import { useI18n } from 'vue-i18n'
 
-import LoadingSpinner from '@/components/common/LoadingSpinner.vue'
+import Loader from '@/components/common/Loader.vue'
 import ScrubableNumberInput from '@/components/common/ScrubableNumberInput.vue'
 import Popover from '@/components/ui/Popover.vue'
 import Button from '@/components/ui/button/Button.vue'
@@ -267,7 +267,7 @@ defineExpose({ runButtonClick })
             </Button>
           </template>
           <template v-else>
-            <LoadingSpinner size="sm" />
+            <Loader size="sm" />
             <span v-text="t('queue.jobQueueing')" />
           </template>
         </div>
